==> scripts/04_analyses_controlling-for-sub-regions.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}

library(tidyverse)
library(haven)
library(boot)

```

RESULTS CONTROLLING FOR SUB REGIONS:

# Read data:
```{r}

# Data-set containing 22 variables from the World Bank, the United Nations, and Gallup:
data_combined <- read_rds("../data-processed/full-data-set_transformed-variables.rds")

```

# Prepare data-set suitable for upcoming analyses:
```{r}

# Data-set containing all combinations of variables:
data_clean <- data_combined %>% 
  distinct(var) %>% 
  rename(var1 = var) %>% 
  mutate(var2 = var1) %>% 
  complete(var1, var2) %>% 
  filter(var1 != var2)

# Add data pertaining to the variables:
data_clean <- data_clean %>% 
  left_join(data_combined, by = c("var1" = "var")) %>% 
  rename(data1 = value) %>% 
  left_join(data_combined %>% 
              select(-region, -sub_region), by = c("country", "var2" = "var")) %>% 
  rename(data2 = value) 

# Nest:
data_nested <- data_clean %>% 
  rename(outcome_var = var1, outcome_value = data1, predictor_var = var2, predictor_value = data2) %>% 
  group_by(outcome_var, predictor_var) %>% 
  nest()

```

# Analyses:

## Full data-set:

For each pair outcome-predictor, bootstrap b-values and confidence intervals:
```{r}

# Function to bootstrap:
boot_slope_ctrl <- function(data,i) {
  
  # Run linear model:
  m <- lm(data$outcome_value[i] ~ data$predictor_value[i] + data$sub_region[i])
  
  # Retrieve b-value:
  summary <- summary.lm(m)
  summary$coefficients[2]
  
}

# Bootstrap b-values and confidence intervals: This takes half an hour to run!
# boot_bvalues_ctrl <- data_nested %>%
#   mutate(boot_b = map(data, ~boot(., boot_slope_ctrl, 2000)),
#          b_value = map(boot_b, ~unlist(.[[1]][1])),
#          boot_ci = map(boot_b, ~boot.ci(.)),
#          ci = map(boot_ci, ~.$normal))
# 
# write_rds(boot_bvalues_ctrl, "../data-processed/bootstrapped-b-values_full-data_controlling-for-subregion.rds")
boot_bvalues_ctrl <- read_rds("../data-processed/bootstrapped-b-values_full-data_controlling-for-subregion.rds")

bvalues_ctrl <- boot_bvalues_ctrl %>% 
  select(-data, -boot_b, -boot_ci, -ci) %>% 
  unnest(b_value) %>% 
  rename(true_b_value = b_value)

ci <- boot_bvalues_ctrl %>% 
  select(-data, -boot_b, -boot_ci, -b_value) %>% 
  mutate(lower_boundary = map(ci, ~.[[2]]),
         upper_boundary = map(ci, ~.[[3]])) %>% 
  unnest(c(lower_boundary, upper_boundary))

```

## Sub-samples of countries:

How often is the true slope within the sub-sample confidence interval?

### Internation Study of Metanorms (ISMN) - 2019 data:
```{r}

ismn_countries <- read_rds("../data-raw/ISMN_list-of-countries.rds")

ismn_data <- data_nested %>% 
  unnest() %>% 
  filter(country %in% ismn_countries) %>% 
  group_by(outcome_var, predictor_var) %>% 
  nest() 

# # Bootstrap b-values and confidence intervals:
# ismn_boot_ctrl <- ismn_data %>%
#   mutate(boot_b = map(data, ~boot(., boot_slope_ctrl, 2000)),
#          b_value = map(boot_b, ~unlist(.[[1]][1])),
#          boot_ci = map(boot_b, ~boot.ci(.)),
#          ci = map(boot_ci, ~.$normal))
# 
# write_rds(ismn_boot_ctrl, "../data-processed/bootstrapped-b-values_ISMN-data_controlling-for-subregion.rds")
ismn_boot_ctrl <- read_rds("../data-processed/bootstrapped-b-values_ISMN-data_controlling-for-subregion.rds")

# Isolate confidence intervals:
ismn_ci <- ismn_boot_ctrl %>% 
  select(-data, -boot_b, -boot_ci, -b_value) %>% 
  mutate(lower_boundary = map(ci, ~.[[2]]),
         upper_boundary = map(ci, ~.[[3]])) %>% 
  unnest(c(lower_boundary, upper_boundary)) %>% 
  select(-ci)

# How often is the true slope within the sub-sample confidence interval?
ismn_result_ctrl <- ismn_ci %>% 
  left_join(bvalues_ctrl) %>% 
  mutate(within_ci = ifelse(true_b_value >= lower_boundary & true_b_value <= upper_boundary, 1, 0)) %>% 
  ungroup() %>% 
  summarise(mean(within_ci))

```

### World Values Survey (WVS) - Wave 7:
```{r}

wvs_countries <- read_rds("../data-raw/WVS_list-of-countries.rds")

wvs_data <- data_nested %>% 
  unnest() %>% 
  filter(country %in% wvs_countries) %>% 
  group_by(outcome_var, predictor_var) %>% 
  nest() 

# # Bootstrap b-values and confidence intervals:
# wvs_boot_ctrl <- wvs_data %>%
#   mutate(boot_b = map(data, ~boot(., boot_slope_ctrl, 2000)),
#          b_value = map(boot_b, ~unlist(.[[1]][1])),
#          boot_ci = map(boot_b, ~boot.ci(.)),
#          ci = map(boot_ci, ~.$normal))
# 
# write_rds(wvs_boot_ctrl, "../data-processed/bootstrapped-b-values_WVS-data_controlling-for-subregion.rds")
wvs_boot_ctrl <- read_rds("../data-processed/bootstrapped-b-values_WVS-data_controlling-for-subregion.rds")

# Isolate confidence intervals:
wvs_ci <- wvs_boot_ctrl %>% 
  select(-data, -boot_b, -boot_ci, -b_value) %>% 
  mutate(lower_boundary = map(ci, ~.[[2]]),
         upper_boundary = map(ci, ~.[[3]])) %>% 
  unnest(c(lower_boundary, upper_boundary)) %>% 
  select(-ci)

# How often is the true slope within the sub-sample confidence interval?
wvs_result_ctrl <- wvs_ci %>% 
  left_join(bvalues_ctrl) %>% 
  mutate(within_ci = ifelse(true_b_value >= lower_boundary & true_b_value <= upper_boundary, 1, 0)) %>% 
  ungroup() %>% 
  summarise(mean(within_ci))

```

### International Social Survey Programme (ISSP) - 2018 data:
```{r}

issp_countries <- read_rds("../data-raw/ISSP_list-of-countries.rds")

issp_data <- data_nested %>% 
  unnest() %>% 
  filter(country %in% issp_countries) %>% 
  group_by(outcome_var, predictor_var) %>% 
  nest() 

# # Bootstrap b-values and confidence intervals:
# issp_boot_ctrl <- issp_data %>%
#   mutate(boot_b = map(data, ~boot(., boot_slope_ctrl, 2000)),
#          b_value = map(boot_b, ~unlist(.[[1]][1])),
#          boot_ci = map(boot_b, ~boot.ci(.)),
#          ci = map(boot_ci, ~.$normal))
# 
# write_rds(issp_boot_ctrl, "../data-processed/bootstrapped-b-values_ISSP-data_controlling-for-subregion.rds")
issp_boot_ctrl <- read_rds("../data-processed/bootstrapped-b-values_ISSP-data_controlling-for-subregion.rds")

# Isolate confidence intervals:
issp_ci <- issp_boot_ctrl %>% 
  select(-data, -boot_b, -boot_ci, -b_value) %>% 
  mutate(lower_boundary = map(ci, ~.[[2]]),
         upper_boundary = map(ci, ~.[[3]])) %>% 
  unnest(c(lower_boundary, upper_boundary)) %>% 
  select(-ci)

# How often is the true slope within the sub-sample confidence interval?
issp_result_ctrl <- issp_ci %>% 
  left_join(bvalues_ctrl) %>% 
  mutate(within_ci = ifelse(true_b_value >= lower_boundary & true_b_value <= upper_boundary, 1, 0)) %>% 
  ungroup() %>% 
  summarise(mean(within_ci))

```

### Programme for International Student Assessment (PISA) - 2015 data:
```{r}

pisa_countries <- read_rds("../data-raw/PISA_list-of-countries.rds")

pisa_data <- data_nested %>% 
  unnest() %>% 
  filter(country %in% pisa_countries) %>% 
  group_by(outcome_var, predictor_var) %>% 
  nest() 

# # Bootstrap b-values and confidence intervals:
# pisa_boot_ctrl <- pisa_data %>%
#   mutate(boot_b = map(data, ~boot(., boot_slope_ctrl, 2000)),
#          b_value = map(boot_b, ~unlist(.[[1]][1])),
#          boot_ci = map(boot_b, ~boot.ci(.)),
#          ci = map(boot_ci, ~.$normal))
# 
# write_rds(pisa_boot_ctrl, "../data-processed/bootstrapped-b-values_PISA-data_controlling-for-subregion.rds")
pisa_boot_ctrl <- read_rds("../data-processed/bootstrapped-b-values_PISA-data_controlling-for-subregion.rds")

# Isolate confidence intervals:
pisa_ci <- pisa_boot_ctrl %>% 
  select(-data, -boot_b, -boot_ci, -b_value) %>% 
  mutate(lower_boundary = map(ci, ~.[[2]]),
         upper_boundary = map(ci, ~.[[3]])) %>% 
  unnest(c(lower_boundary, upper_boundary)) %>% 
  select(-ci)

# How often is the true slope within the sub-sample confidence interval?
pisa_result_ctrl <- pisa_ci %>% 
  left_join(bvalues_ctrl) %>% 
  mutate(within_ci = ifelse(true_b_value >= lower_boundary & true_b_value <= upper_boundary, 1, 0)) %>% 
  ungroup() %>% 
  summarise(mean(within_ci))

```

### Trends in International Mathematics and Science Study (TIMSS) - 2015 data:
```{r}

timss_countries <- read_rds("../data-raw/TIMSS_list-of-countries.rds")

timss_data <- data_nested %>% 
  unnest() %>% 
  filter(country %in% timss_countries) %>% 
  group_by(outcome_var, predictor_var) %>% 
  nest() 

# # Bootstrap b-values and confidence intervals:
# timss_boot_ctrl <- timss_data %>%
#   mutate(boot_b = map(data, ~boot(., boot_slope_ctrl, 2000)),
#          b_value = map(boot_b, ~unlist(.[[1]][1])),
#          boot_ci = map(boot_b, ~boot.ci(.)),
#          ci = map(boot_ci, ~.$normal))
# 
# write_rds(timss_boot_ctrl, "../data-processed/bootstrapped-b-values_TIMSS-data_controlling-for-subregion.rds")
timss_boot_ctrl <- read_rds("../data-processed/bootstrapped-b-values_TIMSS-data_controlling-for-subregion.rds")

# Isolate confidence intervals:
timss_ci <- timss_boot_ctrl %>% 
  select(-data, -boot_b, -boot_ci, -b_value) %>% 
  mutate(lower_boundary = map(ci, ~.[[2]]),
         upper_boundary = map(ci, ~.[[3]])) %>% 
  unnest(c(lower_boundary, upper_boundary)) %>% 
  select(-ci)

# How often is the true slope within the sub-sample confidence interval?
timss_result_ctrl <- timss_ci %>% 
  left_join(bvalues_ctrl) %>% 
  mutate(within_ci = ifelse(true_b_value >= lower_boundary & true_b_value <= upper_boundary, 1, 0)) %>% 
  ungroup() %>% 
  summarise(mean(within_ci))

```

# Figures:

## Figure 1:
```{r}

results_combined <- read_rds("../data-processed/results-combined.rds")

results_combined %>% 
# Add results controlled for world region to data_combined:
  bind_rows(ismn_result_ctrl %>% mutate(source = "ISMN", ctrl = "Controlled"),
            wvs_result_ctrl %>% mutate(source = "WVS", ctrl = "Controlled"),
            issp_result_ctrl %>% mutate(source = "ISSP", ctrl = "Controlled"),
            pisa_result_ctrl %>% mutate(source = "PISA", ctrl = "Controlled"),
            timss_result_ctrl %>% mutate(source = "TIMSS", ctrl = "Controlled")) %>%
  mutate(ctrl = ifelse(is.na(ctrl), "Not controlled", ctrl)) %>% 
  ggplot(aes(x = source, y = `mean(within_ci)`, fill = factor(ctrl, levels = c("Not controlled", "Controlled")))) +
  geom_col(color = "black", position = position_dodge2(width = 0.5)) +
  scale_fill_viridis_d(begin = 0.15, end = 0.75) +
  ylim(0, 1) +
  theme(legend.title = element_blank()) +
  labs(x = NULL, y = NULL) +
  ggtitle("How often is the true regression slope\nwithin bootstrapped 95% confidence intervals?")

ggsave("../figures/main-results.jpeg", width = 10, height = 5.5)

```

Summary table:
```{r}

data_combined %>% 
  bind_rows(ismn_result_ctrl %>% mutate(source = "ISMN", ctrl = "Controlled"),
            wvs_result_ctrl %>% mutate(source = "WVS", ctrl = "Controlled"),
            issp_result_ctrl %>% mutate(source = "ISSP", ctrl = "Controlled"),
            pisa_result_ctrl %>% mutate(source = "PISA", ctrl = "Controlled"),
            timss_result_ctrl %>% mutate(source = "TIMSS", ctrl = "Controlled")) %>% 
  mutate(ctrl = ifelse(is.na(ctrl), "Not controlled", ctrl)) %>% 
  spread(ctrl, `mean(within_ci)`) %>% 
  select(Source = source, `Not controlled`, `Controlled`) %>% 
  mutate_at(vars(`Not controlled`, `Controlled`), ~round(., 2)) %>% 
  kableExtra::kbl()

```

## Figure 2:

Histograms of HDI scores for:

1. The entire set of the world's countries:
```{r}

hdi_2021data <- readxl::read_xlsx("../data-raw/HDI_21-22.xlsx")

# Clean data-set:
hdi_2021data <- hdi_2021data %>% 
  slice(-1:-7) %>% 
  rename(`HDI rank` = Back, 
         Country = `Table 1. Human Development Index and its components`,
         HDI = `...3`,
         `Life expectancy at birth` = `...5`,
         `Expected years of schooling` = `...7`,
         `Mean years of schooling` = `...9`,
         `Gross national income (GNI) per capita` = `...11`) %>% 
  # Keep relevant variables only:
  select(`HDI rank`:HDI, `Life expectancy at birth`, `Expected years of schooling`, `Mean years of schooling`, `Gross national income (GNI) per capita`) %>% 
  mutate_at(vars(HDI:`Gross national income (GNI) per capita`), ~as.numeric(.)) %>% 
  drop_na(`HDI rank`, HDI) %>% 
  # Clean country names:
  mutate(Country = case_when(
    Country == "Bolivia (Plurinational State of)" ~ "Bolivia", 
    Country == "Congo (Democratic Republic of the)" ~ "Congo, Democratic Republic of the", 
    Country == "Côte d'Ivoire" ~ "Cote d'Ivoire", 
    Country == "Czechia" ~ "Czech Republic", 
    Country == "Eswatini (Kingdom of)" ~ "Eswatini", 
    Country == "Hong Kong, China (SAR)" ~ "Hong Kong", 
    Country == "Korea (Republic of)" ~ "South Korea", 
    Country == "Moldova (Republic of)" ~ "Moldova, Republic of", 
    Country == "Tanzania (United Republic of)" ~ "Tanzania, United Republic of", 
    Country == "Türkiye" ~ "Turkey", 
    Country == "Venezuela (Bolivarian Republic of)" ~ "Venezuela", 
    TRUE ~ Country))

```

```{r}

# Histogram:
(panel_a <- hdi_2021data %>% 
  ggplot(aes(x = HDI)) +
  geom_histogram(fill = "midnightblue", color = "white") +
  labs(y = NULL, x = NULL) +
  ggtitle("191 countries"))
  
```

2. Only ISMN countries:
```{r}

(panel_b <- hdi_2021data %>% 
  filter(Country %in% ismn_countries) %>% 
  ggplot(aes(x = HDI)) +
  geom_histogram(fill = "midnightblue", color = "white") +
  labs(y = NULL, x = NULL) +
  ggtitle("57 ISMN countries"))

```

3. Only WVS countries:
```{r}

(panel_c <- hdi_2021data %>% 
  filter(Country %in% wvs_countries) %>% 
  ggplot(aes(x = HDI)) +
  geom_histogram(fill = "midnightblue", color = "white") +
  labs(y = NULL, x = NULL) +
  ggtitle("46 WVS countries"))

```

4. Only ISSP countries:
```{r}

(panel_d <- hdi_2021data %>% 
  filter(Country %in% issp_countries) %>% 
  ggplot(aes(x = HDI)) +
  geom_histogram(fill = "midnightblue", color = "white") +
  labs(y = NULL, x = NULL) +
  ggtitle("32 ISSP countries"))

```

5. Only PISA countries:
```{r}

(panel_e <- hdi_2021data %>% 
  filter(Country %in% pisa_countries) %>% 
  ggplot(aes(x = HDI)) +
  geom_histogram(fill = "midnightblue", color = "white") +
  labs(y = NULL, x = NULL) +
  ggtitle("65 PISA countries"))

```

6. Only TIMSS countries:
```{r}

(panel_f <- hdi_2021data %>% 
  filter(Country %in% timss_countries) %>% 
  ggplot(aes(x = HDI)) +
  geom_histogram(fill = "midnightblue", color = "white") +
  labs(y = NULL, x = "HDI score") +
  ggtitle("36 TIMSS countries"))

```

All panels:
```{r}

cowplot::plot_grid(panel_a, panel_b, panel_c, panel_d, panel_e, panel_f, align = "hv", ncol = 1)

ggsave("../figures/figure2_histograms.jpeg", width = 6.5, height = 15)

```

## Figure 3:
```{r}

bvalues %>%
  rename(bvalue = true_b_value) %>% 
  left_join(bvalues_ctrl %>% 
              rename(bvalue_ctrl = true_b_value)) %>% 
  mutate(quotient = bvalue_ctrl/bvalue) %>% 
  ggplot(aes(x = quotient)) +
  geom_histogram(fill = "lightgrey", color = "black", binwidth = 1) +
  labs(x = "Quotient", y = NULL)

ggsave("../figures/figure3_quotients.jpeg", width = 6.5, height = 4.5)

```
